Fill majority script (WIP) + eval settings
